backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai==1.51.0
python-dotenv==1.0.0
httpx==0.27.0
ultralytics==8.0.196
pillow==10.1.0
opencv-python-headless==4.8.1.78
gunicorn==21.2.0